Move Session abstraction to consulutil
