+++ pkg/kp/consulutil/session_controller_test.go
